projects-display-dev-1 | Re-usable component / Refactoring
Refactored the Project display component to be moved from the home-components folder to now live in it's own component.

Since the only difference between the main projects display component and the one that lives in the home page is the css styling. I decided to make it a reusable component.

This way, there much less code actually written and it's much more functional and managable. I plan on refactoring the ProjectsDisplay component to use classNames and keys passed down via props rather than hard coded that way we can still still have separate CSS styling applied to the two different renders.
