+++ client/src/components/home/home.jsx
@@ -5,14 +5,14 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import "../../dist/styles.css";
 // home page child components
 import AboutMeHome from "./home-components/about-me/about-me.jsx";
-import ProjectsDisplay from "./home-components/projects-display/projects-display.jsx";
+import ProjectsDisplay from "../projects-display/projects-display.jsx";
 
 // home page parent component
 function Home () {
   return (
     <Container key="home-parent" className="home-parent">
       <AboutMeHome key="about-me-home"/>,
-      <ProjectsDisplay key="projects-diplay"/>
+      <ProjectsDisplay key="projects-display"/>
     </Container>
   )
 }
